fix(present): Token-Handoff aus dem Initial-Bundle halten
Tablets öffnen die Presenter-Ansicht über die Startseite, damit der Host-Guard
das Handoff nicht ins Hauptbundle zieht und das Produktionsbudget hält.
