+++ apps/frontend/src/app/core/host-session-token.ts
@@ -1,14 +1,6 @@
 const HOST_TOKEN_STORAGE_PREFIX = 'arsnova-host-token:';
-const HOST_TOKEN_HANDOFF_KEY = 'arsnova-host-token-handoff';
-const HOST_TOKEN_HANDOFF_TTL_MS = 30_000;
 const hostTokens = new Map<string, string>();
 let hostTokensLoaded = false;
-
-type HostTokenHandoff = {
-  code: string;
-  token: string;
-  expiresAt: number;
-};
 
 function isBrowser(): boolean {
   return globalThis.window !== undefined;
@@ -20,59 +12,6 @@ export function normalizeHostSessionCode(sessionCode: string): string {
 
 function getHostTokenStorageKey(sessionCode: string): string {
   return `${HOST_TOKEN_STORAGE_PREFIX}${normalizeHostSessionCode(sessionCode)}`;
-}
-
-function readHostTokenHandoff(): HostTokenHandoff | null {
-  if (!isBrowser()) {
-    return null;
-  }
-  let raw: string | null;
-  try {
-    raw = globalThis.window.localStorage.getItem(HOST_TOKEN_HANDOFF_KEY);
-  } catch {
-    return null;
-  }
-  if (!raw) {
-    return null;
-  }
-  try {
-    globalThis.window.localStorage.removeItem(HOST_TOKEN_HANDOFF_KEY);
-  } catch {
-    // Private mode: still try to parse the in-memory snapshot.
-  }
-  try {
-    const parsed = JSON.parse(raw) as Partial<HostTokenHandoff>;
-    if (
-      typeof parsed.code !== 'string' ||
-      typeof parsed.token !== 'string' ||
-      typeof parsed.expiresAt !== 'number' ||
-      !parsed.code.trim() ||
-      !parsed.token.trim() ||
-      parsed.expiresAt < Date.now()
-    ) {
-      return null;
-    }
-    return {
-      code: normalizeHostSessionCode(parsed.code),
-      token: parsed.token.trim(),
-      expiresAt: parsed.expiresAt,
-    };
-  } catch {
-    return null;
-  }
-}
-
-function consumeHostTokenHandoff(): string | null {
-  const handoff = readHostTokenHandoff();
-  if (!handoff) {
-    return null;
-  }
-  setHostToken(handoff.code, handoff.token);
-  return handoff.code;
-}
-
-export function takeHostTokenHandoffSessionCode(): string | null {
-  return consumeHostTokenHandoff();
 }
 
 function loadHostTokensFromSessionStorage(): void {
@@ -98,22 +37,11 @@ function loadHostTokensFromSessionStorage(): void {
 
 export function getHostToken(sessionCode: string): string | null {
   loadHostTokensFromSessionStorage();
-  const normalizedSessionCode = normalizeHostSessionCode(sessionCode);
-  const existing = hostTokens.get(normalizedSessionCode);
-  if (existing) {
-    return existing;
-  }
-  consumeHostTokenHandoff();
-  return hostTokens.get(normalizedSessionCode) ?? null;
+  return hostTokens.get(normalizeHostSessionCode(sessionCode)) ?? null;
 }
 
 export function hasHostToken(sessionCode: string): boolean {
   return getHostToken(sessionCode) !== null;
-}
-
-export function hasAnyHostToken(): boolean {
-  loadHostTokensFromSessionStorage();
-  return hostTokens.size > 0;
 }
 
 export function getSessionEntryCommands(sessionCode: string): string[] {
@@ -147,38 +75,4 @@ export function setHostToken(sessionCode: string, token: string | null): void {
 
 export function clearHostToken(sessionCode: string): void {
   setHostToken(sessionCode, null);
-}
-
-/** Kurzlebige Token-Übergabe in denselben Origin, weil Tablets sessionStorage nicht in den neuen Tab klonen. */
-export function stageHostTokenHandoff(sessionCode: string): void {
-  if (!isBrowser()) {
-    return;
-  }
-  const token = getHostToken(sessionCode);
-  if (!token) {
-    return;
-  }
-  const payload: HostTokenHandoff = {
-    code: normalizeHostSessionCode(sessionCode),
-    token,
-    expiresAt: Date.now() + HOST_TOKEN_HANDOFF_TTL_MS,
-  };
-  try {
-    globalThis.window.localStorage.setItem(HOST_TOKEN_HANDOFF_KEY, JSON.stringify(payload));
-  } catch {
-    // Private mode / quota: Presenter versucht danach den direkten sessionStorage-Copy.
-  }
-}
-
-export function copyHostTokenToSessionStorage(storage: Storage, sessionCode: string): boolean {
-  const token = getHostToken(sessionCode);
-  if (!token) {
-    return false;
-  }
-  try {
-    storage.setItem(getHostTokenStorageKey(sessionCode), token);
-    return true;
-  } catch {
-    return false;
-  }
 }
