+++ docs/GLOSSAR.md
@@ -44,7 +44,7 @@
 - **Motivationsmeldungen:** Kontextbezogene Rückmeldungen, die Teilnehmende nach Antworten oder Ergebnissen ermutigen bzw. einordnen; Teil des Motivations- und Feedback-Designs. Vertiefung: Story 5.7.
 - **Numerische Schätzfrage / `NUMERIC_ESTIMATE`:** Bewertbarer Fragetyp für Zahlen-Schätzungen mit Zahlentyp, Referenzwert und Toleranzband. Unterstützt optional zwei Runden mit Diskussionsphase sowie beamer-taugliche Statistik und Histogramm-Auswertung nach der Freigabe. Vertiefung: Story 1.2d, `docs/features/numeric-estimate.md`.
 - **Peer Instruction:** Zweite Abstimmungsrunde **im Quiz** (Vorher/Nachher), unabhängig von Blitzlicht-Vergleichsrunde. Vorkommen: Session `currentRound`. Vertiefung: Story 2.7.
-- **Presenter / Beamer:** Projektionsansicht für den Hörsaal ohne Host-Steuerung. Zentrierte Beamer-Bühne (Lobby max. 40rem, Quiz max. 76rem) mit hörsaaltauglicher Typografie; alle Vote-Frageinhalte außer Avatar, Eingaben und Teilnehmer-Chrome werden mitprojiziert, ohne die Fläche auf volle HDMI-Breite zu strecken. Zeigt Lobby (Code/QR), aktuelle Frage, Lesephase, Abstimmung, Ergebnisse, Wortwolke, Q&A und Blitzlicht. Die Lobby füllt den Viewport ohne Scroll: ohne Teams und ohne Motiv ein zentriertes Beitrittsplakat (Code/QR); mit Publikum auf Tablets (Hoch- und Querformat, inkl. kurzer Viewports wie Galaxy Tab S4 1138×712) zwei Spalten Beitritt|Publikum, auf Beamer drei Spalten; Team- und Personenbadges verdichten sich (bei vielen nur Icon/Sequenz), damit alle Teilnehmenden ohne Scroll sichtbar bleiben. Der Host-Button „Presenter-Ansicht“ erscheint auf Desktop und Tablets (iPad, Galaxy Tab), nicht auf schmalen Smartphones. Der neue Tab bekommt eine absolute Presenter-URL; auf Touch-Geräten als `_blank`, plus kurzlebige same-origin Token-Übergabe, weil Tablets `sessionStorage` nicht mit `window.open` klonen. Alternativ den Host-Tab duplizieren und `/host` durch `/present` ersetzen. Vorkommen: Route `session/.../present`. Vertiefung: Story 2.5.
+- **Presenter / Beamer:** Projektionsansicht für den Hörsaal ohne Host-Steuerung. Zentrierte Beamer-Bühne (Lobby max. 40rem, Quiz max. 76rem) mit hörsaaltauglicher Typografie; alle Vote-Frageinhalte außer Avatar, Eingaben und Teilnehmer-Chrome werden mitprojiziert, ohne die Fläche auf volle HDMI-Breite zu strecken. Zeigt Lobby (Code/QR), aktuelle Frage, Lesephase, Abstimmung, Ergebnisse, Wortwolke, Q&A und Blitzlicht. Die Lobby füllt den Viewport ohne Scroll: ohne Teams und ohne Motiv ein zentriertes Beitrittsplakat (Code/QR); mit Publikum auf Tablets (Hoch- und Querformat, inkl. kurzer Viewports wie Galaxy Tab S4 1138×712) zwei Spalten Beitritt|Publikum, auf Beamer drei Spalten; Team- und Personenbadges verdichten sich (bei vielen nur Icon/Sequenz), damit alle Teilnehmenden ohne Scroll sichtbar bleiben. Der Host-Button „Presenter-Ansicht“ erscheint auf Desktop und Tablets (iPad, Galaxy Tab), nicht auf schmalen Smartphones. Der neue Tab bekommt auf Desktop eine absolute Presenter-URL; auf Touch-Geräten `_blank` zur Startseite plus kurzlebige same-origin Token-Übergabe, weil Tablets `sessionStorage` nicht mit `window.open` klonen — Home übernimmt das Token und wechselt in die Presenter-Ansicht. Alternativ den Host-Tab duplizieren und `/host` durch `/present` ersetzen. Vorkommen: Route `session/.../present`. Vertiefung: Story 2.5.
 - **Preset:** Voreinstellung **Seriös** oder **Spielerisch**. Als Quiz-Preset setzt es Defaults für Gamification, Zeit, Lesephase, **Action Sounds** usw.; als UI-Preset ist es eine lokale Browser-Auswahl. In Live-Sessions überschreibt der Host damit nicht das Preset oder Theme von Join-, Vote- oder Present-Clients. Vorkommen: Quiz neu/bearbeiten, Startseite, Header-Toggle. Kanonische Quiz-Defaults: `QUIZ_PRESETS` in shared-types. Vertiefung: [preset-modes](features/preset-modes.md).
 - **Preset-Toast:** Overlay zum Feintuning der Preset-Optionen und Speichern in `localStorage`. Vorkommen: Startseite. Vertiefung: `PresetToastComponent`.
 - **Q&A-Kanal:** Fragen einreichen, Up-/Downvotes, Moderation durch Host und Sortierung nach `Top`, `Beste Fragen` oder `Umstritten`. Vorkommen: Session-Tab. Vertiefung: Epic 8, ADR-0009, [controversy-score](features/controversy-score.md).
